更改了生命周期的基类
