+++ lib/page_life_cycle.dart
@@ -6,7 +6,6 @@ class PageLifeCycleObserver extends WidgetsBindingObserver {
   // route name
   List<int> appRouteNames = [];
 
-
   // 生命周期监控
   List<PageLifeCycle> _observers = [];
   // 缓存上一次的app状态
@@ -26,7 +25,6 @@ class PageLifeCycleObserver extends WidgetsBindingObserver {
 
   void addRouteName(int route) {
     if (appRouteNames.isNotEmpty) {
-
       int hideName = appRouteNames.last;
       _observers
           .where((PageLifeCycle item) => item._pageRouteHashCode == hideName)
@@ -34,10 +32,8 @@ class PageLifeCycleObserver extends WidgetsBindingObserver {
     }
     appRouteNames.add(route);
   }
-  
 
   void removeRouteName(int route, {String type}) {
-
     if (type == 'remove') {
       var index = appRouteNames.lastIndexWhere((int v) => v == route);
       if (index != -1) {
@@ -47,15 +43,16 @@ class PageLifeCycleObserver extends WidgetsBindingObserver {
     if (appRouteNames.isNotEmpty) {
       int showName = appRouteNames.last;
       Future.delayed(Duration.zero).then((_) {
-      if (showName != null) {
-        _observers
-            .where((PageLifeCycle item) => item._pageRouteHashCode == showName)
-            ?.forEach((PageLifeCycle item) {
-              if(appRouteNames.lastIndexOf(item._pageRouteHashCode) != -1) {
-                item?.onShow();
-              }
-              });
-      }
+        if (showName != null) {
+          _observers
+              .where(
+                  (PageLifeCycle item) => item._pageRouteHashCode == showName)
+              ?.forEach((PageLifeCycle item) {
+            if (appRouteNames.lastIndexOf(item._pageRouteHashCode) != -1) {
+              item?.onShow();
+            }
+          });
+        }
       });
     }
   }
@@ -68,7 +65,7 @@ class PageLifeCycleObserver extends WidgetsBindingObserver {
       throw ArgumentError('Route name not is null');
 
     observer._pageRouteHashCode = route.hashCode;
-    
+
     _observers.add(observer);
   }
 
@@ -96,7 +93,8 @@ class PageLifeCycleObserver extends WidgetsBindingObserver {
           int showName = appRouteNames.last;
           if (showName != null) {
             _observers
-                .where((PageLifeCycle item) => item._pageRouteHashCode == showName)
+                .where(
+                    (PageLifeCycle item) => item._pageRouteHashCode == showName)
                 ?.forEach((PageLifeCycle item) => item.onAppForeground());
           }
         }
@@ -110,8 +108,8 @@ class PageLifeCycleObserver extends WidgetsBindingObserver {
             int showName = appRouteNames.last;
             if (showName != null) {
               _observers
-                  .where(
-                      (PageLifeCycle item) => item._pageRouteHashCode == showName)
+                  .where((PageLifeCycle item) =>
+                      item._pageRouteHashCode == showName)
                   ?.forEach((PageLifeCycle item) => item.onAppBackground());
             }
           }
@@ -140,7 +138,7 @@ class PageNavigatorObserver extends NavigatorObserver {
   @override
   void didPop(Route route, Route previousRoute) {
     super.didPop(route, previousRoute);
-    
+
     String name = route.settings?.name;
 
     if (name != null) {
@@ -175,7 +173,7 @@ class PageNavigatorObserver extends NavigatorObserver {
 //      }
 //      return;
 //    }
-    
+
     if (oldName != null) {
       pageObserver.removeRouteName(oldRoute.hashCode);
     }
@@ -197,7 +195,21 @@ class PageNavigatorObserver extends NavigatorObserver {
 //  }
 }
 
-abstract class PageLifeCycle {
+mixin PageLifeCycle<T extends StatefulWidget> on State<T> {
+  final life = PageLifeCycleObserver();
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    life.addPageLifeCycleObserver(this, ModalRoute.of(context));
+  }
+
+  @override
+  void dispose() {
+    life.removePageLifeCycleObserver(this);
+    super.dispose();
+  }
+
   //不支持没有名称的路由，过滤popupRoute
 
   // 当前widget 所在的路由名称  ModelRoute.of(context).setting.name;
